EVFILT_PROC 정상적으로 동작 및 EV_ERROR의 경우 처리 완료

+++ srcs/server/cycle/Cycle.cpp
@@ -1,5 +1,6 @@
 #include <csignal>
 #include "Cycle.hpp"
+#include "../modules/PidSet.hpp"
 
 std::map<int, Cycle> Cycle::_cycleStorage;
 
@@ -44,10 +45,9 @@ void Cycle::reset()
     _cgiRecvfd = -1;
     _readFile = -1;
     _writeFiles.clear();
-    if (_cgiScriptPid != -1) {
+    if (_cgiScriptPid != -1 && PidSet::found(_cgiScriptPid))
         kill(_cgiScriptPid, SIGKILL);
-        _cgiScriptPid = -1;
-    }
+    _cgiScriptPid = -1;
     _httpResponseHandler.reset();
     _cgiRequestHandler.reset();
     _cgiResponseHandler.reset();
